requirements.txt
hyperon== 0.2.3 
Flask == 3.0.3
Werkzeug==3.0.2
biocypher>=0.5.4
python-dotenv==1.0.1
PyYAML==6.0.1
flask-cors == 4.0.1
Flask-Mail ==0.10.0
openpyxl == 3.1.5
PyJWT == 2.9.0
pymongoose == 1.3.8
openai == 1.51.2
pytest == 8.3.3
gunicorn == 23.0.0
Flask-Limiter == 3.8.0 
tiktoken==0.8.0
httpx==0.27.2
networkx==3.4.2
nanoid==2.0.0
elasticsearch==9.0.1
axiom-py==0.9.1
sentry-sdk==2.43.0
APScheduler==3.11.1
neo4j==6.0.3
celery
accelerate
pysam==0.23.3
redis==4.6.0
fastapi
uvicorn
pydantic-settings
python-multipart
python-socketio